+++ lib/ui/wellness/rings/WellnessRingWidgets.dart
@@ -2,7 +2,9 @@ import 'package:confetti/confetti.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/model/wellness/WellnessReing.dart';
 import 'package:illinois/service/WellnessRings.dart';
+import 'package:illinois/ui/widgets/FavoriteButton.dart';
 import 'package:illinois/utils/AppUtils.dart';
+import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
@@ -297,10 +299,11 @@ class WellnessRingButton extends StatefulWidget{
   final bool showLeftIcon;
   final bool showRightIcon;
   final Color? color;
+  final bool enabled;
   final void Function(BuildContext context) onTapWidget;
   final void Function(BuildContext context)? onTapRightWidget;
 
-  const WellnessRingButton({Key? key, required this.label, this.description, this.showLeftIcon = false, this.showRightIcon = false, this.color, required this.onTapWidget, this.onTapRightWidget}) : super(key: key);
+  const WellnessRingButton({Key? key, required this.label, this.description, this.showLeftIcon = false, this.showRightIcon = false, this.color, required this.onTapWidget, this.onTapRightWidget, this.enabled = true}) : super(key: key);
 
   @override
   State<StatefulWidget> createState() => _WellnessRingButtonState();
@@ -312,7 +315,7 @@ class _WellnessRingButtonState extends State<WellnessRingButton>{
   @override
   Widget build(BuildContext context) {
     return Semantics(label: widget.label, hint: widget.description, button: true, excludeSemantics: true, child:
-    GestureDetector(onTap: () => widget.onTapWidget(context), child:
+    GestureDetector(onTap: () => widget.enabled? widget.onTapWidget(context): null, child:
     Row(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
       Expanded(child:
       Container(decoration: BoxDecoration(color: widget.color ?? Colors.white, borderRadius: BorderRadius.all(Radius.circular(4)), border: Border.all(color: Styles().colors!.surfaceAccent!, width: 1)), child:
@@ -320,10 +323,14 @@ class _WellnessRingButtonState extends State<WellnessRingButton>{
       Row(children: <Widget>[
         widget.showLeftIcon ? Padding(padding: EdgeInsets.only(right: 6), child: _leftIcon) : Container(),
         Expanded(child:
-        Text(widget.label , style: TextStyle(color: widget.color!=null? Colors.white : Styles().colors!.fillColorPrimary!, fontFamily: Styles().fontFamilies!.bold, fontSize: 16), textAlign: TextAlign.start,),
+        Text(widget.label ,
+          style: TextStyle(color:widget.enabled? (widget.color!=null? Colors.white : Styles().colors!.fillColorPrimary!) :  Styles().colors!.disabledTextColor!,
+            fontFamily: Styles().fontFamilies!.bold, fontSize: 16), textAlign: TextAlign.start,),
         ),
         Expanded(child:
-        Text(widget.description ?? "" , style: TextStyle(color: widget.color!=null? Colors.white : Styles().colors!.textSurface!, fontFamily: Styles().fontFamilies!.regular, fontSize: 14), textAlign: TextAlign.end,),
+        Text(widget.description ?? "" ,
+          style: TextStyle(color: widget.enabled? (widget.color!=null? Colors.white : Styles().colors!.textSurface!) : Styles().colors!.disabledTextColor,
+            fontFamily: Styles().fontFamilies!.regular, fontSize: 14), textAlign: TextAlign.end,),
         ),
         widget.showRightIcon ? Padding(padding: EdgeInsets.only(left: 6), child: _rightIcon) : Container(),
       ],),
@@ -338,7 +345,7 @@ class _WellnessRingButtonState extends State<WellnessRingButton>{
   Widget get _leftIcon{
     return Container(
       padding: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
-      child: Image.asset('images/icon-create-event.png', excludeFromSemantics: true, color:  Styles().colors!.fillColorPrimary!),
+      child: Image.asset('images/icon-create-event.png', excludeFromSemantics: true, color: widget.enabled ? Styles().colors!.fillColorPrimary! : Styles().colors!.disabledTextColor),
     ); //TBD
   }
 
@@ -450,5 +457,16 @@ class _AccomplishmentCardState extends State<AccomplishmentCard>{
 
   num _trimDecimal(double value){
     return value % 1 == 0 ? value.toInt() : value;
+  }
+}
+
+class WellnessWidgetHelper{
+  static Widget buildWellnessHeader() {
+    return Row(crossAxisAlignment: CrossAxisAlignment.center, children: [
+      Text(Localization().getStringEx('panel.wellness.ring.create.header.label', 'My Daily Wellness Rings'),
+          overflow: TextOverflow.ellipsis,
+          style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 18, fontFamily: Styles().fontFamilies!.bold)),
+      FavoriteStarIcon(style: FavoriteIconStyle.Button, padding: EdgeInsets.symmetric(horizontal: 16))
+    ]);
   }
 }