+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Wellness Tips loaded from backend [#1833](https://github.com/rokwire/illinois-app/issues/1833).
 - Added some wellness resource items to Wellness Home content selector [#1836](https://github.com/rokwire/illinois-app/issues/1836).
 - Added 8 Dimensions of wellness popup image [#1838](https://github.com/rokwire/illinois-app/issues/1838).
+- ### Changed
+- Wellness Rings updates [#1692](https://github.com/rokwire/illinois-app/issues/1692).
 
 ## [4.0.38] - 2022-06-27
 ### Fixed